# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend Go + Gin untuk aplikasi video conference berbasis LiveKit.\nAuth pakai dua jenis JWT: app JWT (untuk endpoint backend ini) dan LiveKit JWT (untuk connect ke LiveKit server, di-issue oleh /api/token).",
        "title": "Videoconf Backend API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/rooms/{idOrSlug}/lock": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Locked = /api/token reject non-owner. Tidak ngusir participant yang udah connect — kombinasi dengan kick untuk evict semua.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "host-controls"
                ],
                "summary": "Lock room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.lockResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{idOrSlug}/participants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Live state dari LiveKit RoomService — bukan dari DB. Identity tiap participant = app user.id (string).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "host-controls"
                ],
                "summary": "List participant aktif di room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.participantsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "bukan owner",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "LiveKit unreachable",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{idOrSlug}/participants/{identity}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Forced disconnect via LiveKit. Participant bisa rejoin kalau room masih unlocked — kombinasi lock+kick untuk benar-benar exclude.",
                "tags": [
                    "host-controls"
                ],
                "summary": "Kick participant dari room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "identity participant",
                        "name": "identity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{idOrSlug}/participants/{identity}/mute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Backend list track participant, filter by source, terus call LiveKit MutePublishedTrack per track yang match. Source kosong = semua track.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "host-controls"
                ],
                "summary": "Mute / unmute track participant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "identity participant (= app user.id)",
                        "name": "identity",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source: audio|video|empty, muted: bool",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.muteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.muteResponse"
                        }
                    },
                    "400": {
                        "description": "source value invalid",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "bukan owner",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "participant gak connect",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "502": {
                        "description": "LiveKit error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{idOrSlug}/unlock": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner only. Restore akses non-owner ke /api/token (kalau room public).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "host-controls"
                ],
                "summary": "Unlock room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id atau slug",
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.lockResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                }
            }
        },
        "internal_handlers.lockResponse": {
            "type": "object",
            "properties": {
                "is_locked": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_handlers.muteRequest": {
            "type": "object",
            "properties": {
                "muted": {
                    "type": "boolean"
                },
                "source": {
                    "description": "\"audio\" | \"video\" | \"\" (all)",
                    "type": "string"
                }
            }
        },
        "internal_handlers.muteResponse": {
            "type": "object",
            "properties": {
                "muted_tracks": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_handlers.participantsListResponse": {
            "type": "object",
            "properties": {
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "identity": {
                                "type": "string"
                            },
                            "joined_at": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "sid": {
                                "type": "string"
                            },
                            "state": {
                                "type": "string"
                            },
                            "tracks": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "muted": {
                                            "type": "boolean"
                                        },
                                        "sid": {
                                            "type": "string"
                                        },
                                        "source": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" lalu spasi, lalu app JWT yang didapat dari /auth/login atau /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}